+++ src/lib.rs
@@ -71,6 +71,7 @@ pub mod decoder;
 pub mod encoder;
 pub mod error;
 pub mod extrude;
+pub mod graph;
 pub mod instance_bvh;
 pub mod light;
 pub mod material;